requirements.txt
pyyaml==6.0.1
click==8.1.6
selenium==4.11.2
termcolor==2.3.0